+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Field `resouce` has been renamed into `hardware`
 * Complete documentation about hardware booked/bought
 * `mse list` does not required the project name. It's optional.
+* Use `get_server_certificate` from `intel-sgx-ra`
 
 ## \[0.10.2\] - 2023-03-01